add @auth directive

+++ graphql/resolvers/mutations/deleteCourse/index.ts
@@ -3,13 +3,31 @@ import prisma from "../../../../prisma";
 import handleError from "../../../../utils/handleError";
 import inputValidation from "../../../../utils/inputValidation";
 import { DeleteCoursePayload } from "./model";
+import { ResolverContext } from "../../../../models/common.model";
 
 const schema = z.object({
   id: z.string().min(1, "Course id is required."),
 });
 
-const deleteCourse = async (_, payload: DeleteCoursePayload) => {
+const deleteCourse = async (
+  _,
+  payload: DeleteCoursePayload,
+  context: ResolverContext
+) => {
   try {
+    const course = await prisma.course.findUnique({
+      where: {
+        id: payload.id,
+      },
+      select: {
+        authorId: true,
+      },
+    });
+
+    if (!course || course.authorId !== context.user.id) {
+      throw new Error("UnAuthorized: User not allowed to delete this course.");
+    }
+
     inputValidation(schema, payload);
 
     await prisma.course.delete({
